+++ prev/source/js/common/fhir-batch-query.js
@@ -27,6 +27,7 @@ export class FhirBatchQuery {
     batchTimeout = 20
   }) {
     this._serviceBaseUrl = serviceBaseUrl;
+    this._isDbgap = false;
     this._pending = [];
     this._batchTimeoutId = null;
     this._batchTimeout = batchTimeout;
@@ -52,6 +53,13 @@ export class FhirBatchQuery {
    */
   getServiceBaseUrl() {
     return this._serviceBaseUrl;
+  }
+
+  /**
+   * Sets whether the service base url is a dbGap url.
+   */
+  setIsDbgap(value) {
+    this._isDbgap = value;
   }
 
   /**
@@ -97,7 +105,7 @@ export class FhirBatchQuery {
       return this._initializationPromise;
     }
     this._features = {};
-    if (this._serviceBaseUrl === 'https://dbgap-api.ncbi.nlm.nih.gov/fhir/x1') {
+    if (this._isDbgap) {
       this._initializationPromise = new Promise((resolve, _) => {
         Promise.allSettled([
           // Query to extract the consent group that must be included as _security param in particular queries.
@@ -514,7 +522,7 @@ export class FhirBatchQuery {
       }
 
       if (method === 'GET') {
-        sendUrl = this.addParamToUrl(sendUrl, '_type', 'json');
+        sendUrl = this.addParamToUrl(sendUrl, '_format', 'json');
       }
 
       oReq.open(method, sendUrl);